Cast the number to the respective data types for the example.

+++ examples/datatypes.cpp
@@ -72,15 +72,15 @@ int main(int, char **)
         _double *= M_PI;
         _float *= M_PIf;
 
-        _uint8_t += 8u;
-        _uint16_t += 16u;
-        _uint32_t += 32u;
-        _uint64_t += 64UL;
+        _uint8_t += static_cast<uint8_t>(8);
+        _uint16_t += static_cast<uint16_t>(16);
+        _uint32_t += static_cast<uint32_t>(32);
+        _uint64_t += static_cast<uint64_t>(64);
 
-        _int8_t -= 8;
-        _int16_t -= 16;
-        _int32_t -= 32;
-        _int64_t -= 64;
+        _int8_t -= static_cast<int8_t>(8);
+        _int16_t -= static_cast<int16_t>(16);
+        _int32_t -= static_cast<int32_t>(32);
+        _int64_t -= static_cast<int64_t>(64);
 
         sine_wave = offset + amplitude * std::sin(2 * M_PI * frequency * time);
